Move src/amms.py to src/protocols/dexes dir
