+++ Dockerfile.update-amm-pools
@@ -9,6 +9,7 @@ COPY pyproject.toml poetry.lock* ./
 RUN poetry config virtualenvs.create false \
   && poetry install --no-interaction --no-ansi
 
+COPY ./scripts ./scripts
 COPY ./src ./src
 COPY ./db ./db
