+++ scripts/collect_amm_liq.py
@@ -9,7 +9,7 @@ import asyncio
 import sys
 sys.path.append(".")
 
-from src.amms import update_amm_dex_data_continuously # pylint: disable=C0413
+from src.protocols.dexes.amms import update_amm_dex_data_continuously # pylint: disable=C0413
 
 if __name__ == "__main__":
     logging.basicConfig(
